requirements.txt
discord-py-interactions
quart
twilio
